Restructure coral modules
